fix(navbar): set font-weight properly on condition enhancement(irc/channel): better text contrast for disabled channels chore: moved irc panel to its own file
